Fix screensaver unlock for some users
pam_fprintd is very careful to use its own main loop context (as it
should) to prevent getting in the way of the gnome-screensaver.

Unfortunately, there's a small bug in its logic where it ends up
removing a random file descriptor from the default context. In the
bug report that file descriptor is really important.  It's what
tells gnome-screensaver when to drop the dialog.

The patch is as simple as:

-   g_source_remove (timeout_id);
+   g_source_destroy (source);

(g_source_remove always operates on the default context)

https://bugzilla.redhat.com/show_bug.cgi?id=614573